test: correct components location
