+++ src/__tests__/app/components/introspection/enum/IntrospectionEnumTypeComponent.test.tsx
