+++ src/__tests__/app/components/introspection/interface/IntrospectionInterfaceTypeComponent.test.tsx
