+++ src/__tests__/app/components/introspection/shared/FoldableBlockComponent.test.tsx
