+++ src/__tests__/app/components/introspection/shared/IntrospectionTypeComponent.test.tsx
